refactor: use descriptor-based approach, split into different files
