+++ packages/sdk-staking/.papi/descriptors/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!package.json
