+++ packages/sdk-staking/.papi/whitelist.ts
@@ -1,0 +1,7 @@
+import type { DotWhitelistEntry } from "./descriptors"
+
+export const whitelist: DotWhitelistEntry[] = [
+  "query.System.Account",
+  "*.Staking",
+  "const.Balances.ExistentialDeposit",
+]
