+++ packages/sdk-staking/package.json
@@ -11,16 +11,16 @@
     ".": {
       "node": {
         "production": {
-          "import": "./dist/esm/index.mjs",
+          "import": "./dist/esm/src/index.mjs",
           "require": "./dist/min/index.js",
           "default": "./dist/index.js"
         },
-        "import": "./dist/esm/index.mjs",
+        "import": "./dist/esm/src/index.mjs",
         "require": "./dist/index.js",
         "default": "./dist/index.js"
       },
-      "module": "./dist/esm/index.mjs",
-      "import": "./dist/esm/index.mjs",
+      "module": "./dist/esm/src/index.mjs",
+      "import": "./dist/esm/src/index.mjs",
       "require": "./dist/index.js",
       "default": "./dist/index.js"
     },
@@ -34,7 +34,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "tsc --noEmit && rollup -c ../../rollup.config.js",
+    "build": "papi generate --whitelist .papi/whitelist.ts && mv .papi/descriptors/dist/index.mjs .papi/descriptors/dist/index.js && tsc --noEmit && rollup -c ../../rollup.config.js",
     "lint": "prettier --check README.md \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
     "format": "prettier --write README.md \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
     "prepack": "pnpm run build"
