+++ packages/sdk-staking/src/index.ts
@@ -1,3 +1,2 @@
 export { createStakingSdk } from "./staking-sdk"
-export type * from "./descriptors"
 export type * from "./sdk-types"
